+++ expyriment/io/_touchscreen_buttonbox.py
@@ -20,19 +20,28 @@ from _input_output import Input
 class TouchScreenButtonBox(Input):
     """A class implementing a TouchScreenButton."""
 
-    def __init__(self, button_fields):
+    def __init__(self, button_fields,
+                stimuli=[], background_stimulus=None):
         """Initialize a touchscreen button box.
 
         Parameters
         ----------
         button_field : visual Expyriment stimulus or list of stimuli
+            The button fields defines the area on which a click action will be
+            registered. Added field will be numbered starting with 0.
+        stimuli : visual Expyriment stimulus or list of stimuli, optional
+            Additonal visual stimuli that will be presented together with the button
+            fields
+        background_stimulus : visual Expyriment stimulus, optional
+            The background stimulus on which the the touschscreen button fields
+            are presented
 
         Notes
         -----
         Every visual Expyriment stimulus can serve as a touchscreen button field.
         If the TouchScreenButtonBox is presented, it can be checked for events
         using the methods 'check' and 'wait'. Methods return the id of the
-        button field. Added field will be numbered starting with 0.
+        button field.
 
         """
 
@@ -40,10 +49,16 @@ class TouchScreenButtonBox(Input):
 
         if type(button_fields) is not list:
             button_fields = [button_fields]
+        if type(stimuli) is not list:
+            stimuli = [stimuli]
 
-        self._button_fields = []
         self._mouse = None
+        self._canvas = None
+        self._button_fields = []
+        self._stimuli = []
+        self.background_stimulus = background_stimulus
         map(self.add_button_field, button_fields)
+        map(self.add_stimulus, stimuli)
 
     def add_button_field(self, button_field):
         """Add a touchscreen button fields.
@@ -61,48 +76,99 @@ class TouchScreenButtonBox(Input):
         if not isinstance(button_field, expyriment.stimuli._visual.Visual):
             raise TypeError("Button field has to a visual Expyriment stimulus")
         self._button_fields.append(button_field)
+        self._canvas = None
 
+    def add_stimulus(self, stimulus):
+        """Add additonal stimulus that will be presented together with the button
+        fields.
+
+        Parameters
+        ----------
+        stimulus : visual expyriment stimulus
+
+        """
+        if not isinstance(stimulus, expyriment.stimuli._visual.Visual):
+            raise TypeError("Additional stimuli has to a visual Expyriment stimulus")
+        self._stimuli.append(stimulus)
+        self._canvas = None
 
     @property
     def button_field(self):
         """getter of button fields (list of visual Expyriment stimuli)"""
         return self._button_fields
 
-    def clear(self):
+    @property
+    def stimuli(self):
+        """getter of additonal stimuli (list of visual Expyriment stimuli)"""
+        return self._stimuli
+
+    @property
+    def background_stimulus(self):
+        """Getter of background stimulus.
+
+        Background stimulus, on which the button fields and the addtional stimuli
+        will be presented. (visual Expyriment stimuli)
+
+        """
+        return self._background_stimulus
+
+    @background_stimulus.setter
+    def background_stimulus(self, stimulus):
+        """Setter background stimulus"""
+        if stimulus is None:
+            self._background_stimulus = expyriment.stimuli.BlankScreen()
+        elif not isinstance(stimulus, expyriment.stimuli._visual.Visual):
+            raise TypeError("Background stimulus has to be be a " +\
+                            "visual Expyriment stimulus")
+        else:
+            self._background_stimulus = stimulus
+        self._canvas = None
+
+    def clear_event_buffer(self):
         """Clear the event buffer of the touchscreen/mouse input device."""
 
         if self._mouse is not None:
             self._mouse.clear()
 
-    def present(self, background_stimulus=None, show_cursor=True):
+    def preload(self):
+        """Prepare and preload the button fields and addtional stimuli for
+        presentation.
+
+        """
+
+        self._canvas = self._background_stimulus.copy()
+        if len(self._button_fields)<1:
+            raise RuntimeError("No button field defined!")
+        map(lambda x:x.plot(self._canvas), self._button_fields)
+        map(lambda x:x.plot(self._canvas), self._stimuli)
+        self._canvas.preload()
+
+    def unload(self):
+        if self._canvas is not None:
+            self._canvas.unload()
+        self._canvas = None
+
+    def present(self, show_cursor=True):
         """Present touchscreen buttons.
 
         Parameters
         ----------
-        background_stimulus : visual Expyriment stimulus, optional
-            The background stimulus on which the the touschscreen button fields
-            are presented
         show_cursor : bool, optional
             shows mouse cursor (default = True)
 
         """
 
-        if background_stimulus is None:
-            stim = expyriment.stimuli.BlankScreen()
-        elif not isinstance(background_stimulus,
-                                expyriment.stimuli._visual.Visual):
-            raise TypeError("Background stimulus has to be be a " +\
-                            "visual Expyriment stimulus")
+        if self._canvas is None:
+            self.preload()
+        if self._mouse is None:
+            self._mouse = expyriment.io.Mouse()
+            self._mouse.set_logging(False)
+        if show_cursor:
+            self._mouse.show_cursor()
         else:
-            stim = background_stimulus.copy()
-        if len(self._button_fields)<1:
-            raise RuntimeError("No button field defined!")
-
-        map(lambda x:x.plot(stim), self._button_fields)
-        self._mouse = expyriment.io.Mouse(show_cursor=show_cursor)
-        self._mouse.set_logging(False)
+            self._mouse.hide_cursor()
         self._mouse.clear()
-        stim.present()
+        self._canvas.present()
 
     def check(self, button_field_ids=None, check_for_control_keys=True):
         """Check if a button field is clicked.
@@ -189,7 +255,7 @@ class TouchScreenButtonBox(Input):
         """
 
         start = get_time()
-        self.clear()
+        self.clear_event_buffer()
         while True:
             expyriment._active_exp._execute_wait_callback()
             pressed_button = self.check(button_field_ids,