+++ expyriment/control/_experiment_control.py
@@ -10,7 +10,7 @@ __revision__ = ''
 __date__ = ''
 
 import sys
-
+import os
 import pygame
 try:
     import android
@@ -28,7 +28,7 @@ from expyriment.io import DataFile, EventFile, TextInput, Keyboard, Mouse
 from expyriment.io import _keyboard, TouchScreenButtonBox
 from expyriment.io._screen import Screen
 from _miscellaneous import _set_stdout_logging, is_idle_running
-from expyriment.misc import unicode2str
+from expyriment.misc import unicode2str, get_experiment_secure_hash
 
 
 def start(experiment=None, auto_create_subject_id=None, subject_id=None,
@@ -415,10 +415,18 @@ def initialize(experiment=None):
                             text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
                             background_colour=(0, 0, 0),
                             position=(0, 40))
+    text2 = stimuli.TextLine("{0} ({1})".format(
+                            os.path.split(sys.argv[0])[1],
+                            get_experiment_secure_hash()),
+                            text_size=14,
+                            text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
+                            background_colour=(0, 0, 0),
+                            position=(0, 10))
     canvas = stimuli.Canvas((600, 300), colour=(0, 0, 0))
     canvas2 = stimuli.Canvas((600, 300), colour=(0, 0, 0))
     logo.plot(canvas)
     text.plot(canvas)
+    text2.plot(canvas)
     canvas.preload(True)
     canvas._set_surface(canvas._get_surface().convert())
     start = experiment.clock.time