+++ CHANGES.md
@@ -5,12 +5,12 @@ Expyriment Release Notes
 Upcomming Version
 -----------------
 New Features:
-- new io class: TouchScreen Button
+- new io class: TouchScreenButtonBox
 - control.start() can now be called with a predefined subject id
 
 Fixed:
-- shapes can now be used as background stimuli in bug for io.TextInput & 
-  io.TextMenue
+- bugfix: stimuli.shapes can now be used as background stimuli for io.TextInput & 
+  io.TextMenu
 
 Version 0.7.0b1 (15 Feb 2014)
 -----------------------------