+++ CHANGES.md
@@ -9,6 +9,7 @@ New Features:
 - new options for control.start(): skip_ready_screen and subject_id to start with
   predefined subject id
 - experiment have now also a global mouse object: experiment.mouse 
+- new property for io.mouse: is_visible
 
 Fixed:
 - bugfix: stimuli.shapes can now be used as background stimuli for io.TextInput &